text classifier ML Model
Code for text classifying ML model countVectorizer and multinomialNB fronscikit learn to train a Naive Bayes classifier, will integrate sprint 1 code into it later on when we need it, the input has to be manually typed in and cant contain symbol ' in it, something to keep in mind before sending input through frontend. The model is very poorly trained due to insufficient dataset. Put dataset.csv in same folder as the jupyter notebook to run the code